fix: reviewed, closed some threads
